backend/requirements.txt
# FastAPI Framework
fastapi==0.122.0
uvicorn[standard]==0.30.0
pydantic>=2.11.7,<3.0.0
pydantic-settings==2.4.0

# Supabase Client
supabase==2.24.0

# Database ORM
sqlalchemy==2.0.36
asyncpg==0.29.0

# Redis Client
redis==5.0.1

# Celery (for future async tasks)
celery==5.5.3

# LangChain AI Libraries
langchain~=0.3.0
langchain-openai~=0.2.0
langchain-anthropic~=0.3.0

# PDF Processing
pymupdf==1.24.10
ebooklib
Pillow>=10.0.0

# HTTP Client for font downloads
requests>=2.31.0

# Storage (for future file operations)
boto3==1.36.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.9
structlog==24.1.0
python-magic==0.4.27

# Testing
pytest==8.3.0
pytest-asyncio==0.21.2
httpx==0.27.0

# JWT Authentication
python-jose[cryptography]==3.3.0
